Bug 1262945 - Move prelanded/obsolete strings back into Client target
